WRP-17963: [sandstone] Add ui-test for editable scroller (#1510)
* WRP-17963: [sandstone] Increase ui-test for editable scroller
Enact-DCO-1.0-Signed-off-by: Jiye Kim (jiye.kim@lge.com)

* WRP-17963: [sandstone] Increase ui-test for editable scroller
Enact-DCO-1.0-Signed-off-by: Jiye Kim (jiye.kim@lge.com)

* WRP-17963: [sandstone] Increase ui-test for editable scroller
Enact-DCO-1.0-Signed-off-by: Jiye Kim (jiye.kim@lge.com)

* WRP-17963: [sandstone] Increase ui-test for editable scroller
Enact-DCO-1.0-Signed-off-by: Jiye Kim (jiye.kim@lge.com)
