+++ tests/ui/apps/Scroller/WithEditableSelectItemByPress/ScrollerWithEditableSelectItemByPress.module.less
